cloud_functions/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
scikit-learn==1.6.1
